moved "cabinet" component to containers
