+++ src/containers/cabinet/cabinet.styles.js
