+++ src/services/local-storage.service.js
@@ -2,9 +2,9 @@ export const clearLocalStorage = () => {
   const horondi = {
     accessToken: null,
     _id: null,
-    wishlist: null,
+    wishlist: [],
     language: 0,
-    cart: null
+    cart: []
   };
   localStorage.setItem('horondi', JSON.stringify(horondi));
 };
